+++ network/optimizer/amsgrad.py
@@ -11,9 +11,9 @@ class AMSGrad(Optimizer):
 
     def _initialize_param_state(self, param: np.ndarray, name: str, param_id: int):
         if 'm' not in self.param_states[param_id]:
-            self.param_states[param_id]['m'] = np.zeros_like(param, dtype=np.float64)
-            self.param_states[param_id]['v'] = np.zeros_like(param, dtype=np.float64)
-            self.param_states[param_id]['v_max'] = np.zeros_like(param, dtype=np.float64)
+            self.param_states[param_id]['m'] = np.zeros_like(param)
+            self.param_states[param_id]['v'] = np.zeros_like(param)
+            self.param_states[param_id]['v_max'] = np.zeros_like(param)
 
     def _update_single_param(self, param: np.ndarray, grad: np.ndarray, lr: float):
         param_id = id(param)
