+++ network/optimizer/rmsprop.py
@@ -10,7 +10,7 @@ class RMSprop(Optimizer):
 
     def _initialize_param_state(self, param: np.ndarray, name: str, param_id: int):
         if 'square_avg' not in self.param_states[param_id]:
-            self.param_states[param_id]['square_avg'] = np.zeros_like(param, dtype=np.float64)
+            self.param_states[param_id]['square_avg'] = np.zeros_like(param)
 
     def _update_single_param(self, param: np.ndarray, grad: np.ndarray, lr: float):
         param_id = id(param)
